+++ lib/bindings/hci/misc/Signaling.d.ts
@@ -1,10 +1,5 @@
-/// <reference types="node" />
-import { EventEmitter } from 'events';
 import { Hci } from './Hci';
-export declare interface Signaling {
-    on(event: 'connectionParameterUpdateRequest', listener: (minInterval: number, maxInterval: number, latency: number, supervisionTimeout: number) => void): this;
-}
-export declare class Signaling extends EventEmitter {
+export declare class Signaling {
     private hci;
     private handle;
     constructor(hci: Hci, handle: number);