+++ lib/index.d.ts
@@ -1,3 +1,3 @@
-export { Adapter, Noble, Peripheral, Gatt, GattLocal, GattRemote, GattCharacteristic, GattCharacteristicLocal, GattCharacteristicRemote, GattDescriptor, GattDescriptorLocal, GattDescriptorRemote, GattService, GattServiceLocal, GattServiceRemote } from './models';
+export * from './models';
 export { HciNoble } from './bindings/hci';
 export { DbusNoble } from './bindings/dbus';