+++ src/bindings/dbus/Peripheral.ts
@@ -120,7 +120,7 @@ export class DbusPeripheral extends Peripheral {
 		});
 	}
 
-	public async disconnect(): Promise<number> {
+	public async disconnect(): Promise<void> {
 		if (!(await this.isConnected())) {
 			return;
 		}
@@ -130,13 +130,13 @@ export class DbusPeripheral extends Peripheral {
 		}
 
 		if (this.isDisconnecting) {
-			return new Promise<number>((resolve, reject) => this.disconnecting.push([resolve, reject]));
+			return new Promise<void>((resolve, reject) => this.disconnecting.push([resolve, reject]));
 		}
 
 		this.disconnecting = [];
 		this.isDisconnecting = true;
 
-		return new Promise(async (resolve, reject) => {
+		return new Promise<void>(async (resolve, reject) => {
 			this.disconnecting.push([resolve, reject]);
 
 			const done = () => this.doneDisconnecting();